integrated login and signup screens
